# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Controle de Diárias de Prestadores de Serviços
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  gestor: 'gestor',
  prestador: 'prestador',
} as const;

export interface User {
  id: number;
  decargoId: string;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  teamId?: number | null;
  /** @nullable */
  teamName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  active: boolean;
  createdAt: string;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  admin: 'admin',
  gestor: 'gestor',
  prestador: 'prestador',
} as const;

export interface UserUpdate {
  role?: UserUpdateRole;
  /** @nullable */
  teamId?: number | null;
  active?: boolean;
}

export interface Team {
  id: number;
  name: string;
  /** @nullable */
  managerId?: number | null;
  /** @nullable */
  managerName?: string | null;
  providerCount: number;
  createdAt: string;
}

export interface TeamInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  managerId?: number | null;
}

export interface TeamUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  managerId?: number | null;
}

export interface Provider {
  id: number;
  decargoId: string;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  teamId?: number | null;
  /** @nullable */
  teamName?: string | null;
  active: boolean;
  /** @nullable */
  syncedAt?: string | null;
}

export interface SyncResult {
  synced: number;
  created: number;
  updated: number;
  deactivated: number;
}

export type DiariaStatus = typeof DiariaStatus[keyof typeof DiariaStatus];


export const DiariaStatus = {
  pendente_aprovacao: 'pendente_aprovacao',
  em_analise: 'em_analise',
  aprovada: 'aprovada',
  rejeitada: 'rejeitada',
  solicitacao_correcao: 'solicitacao_correcao',
  disponivel_exportacao: 'disponivel_exportacao',
  exportada: 'exportada',
  paga: 'paga',
  cancelada: 'cancelada',
} as const;

export interface Diaria {
  id: number;
  providerId: number;
  providerName: string;
  teamId: number;
  teamName: string;
  /** @nullable */
  managerId?: number | null;
  /** @nullable */
  managerName?: string | null;
  workDate: string;
  /**
     * Null when viewer is a provider (no financial visibility)
     * @nullable
     */
  value?: number | null;
  /** @nullable */
  paymentDate?: string | null;
  /** @nullable */
  observations?: string | null;
  status: DiariaStatus;
  /** @nullable */
  actionNote?: string | null;
  createdAt: string;
  createdByName: string;
  /** @nullable */
  approvedAt?: string | null;
  /** @nullable */
  approvedByName?: string | null;
  /** @nullable */
  exportedAt?: string | null;
  /** @nullable */
  exportedByName?: string | null;
  /** @nullable */
  integrationId?: string | null;
  /** @nullable */
  paidAt?: string | null;
  /** @nullable */
  cancelledAt?: string | null;
}

export interface DiariaInput {
  providerId: number;
  teamId: number;
  workDate: string;
  value: number;
  /** @nullable */
  paymentDate?: string | null;
  /** @nullable */
  observations?: string | null;
}

export interface DiariaUpdate {
  workDate?: string;
  value?: number;
  /** @nullable */
  paymentDate?: string | null;
  /** @nullable */
  observations?: string | null;
}

export interface DiariaActionNote {
  /** @nullable */
  note?: string | null;
}

export interface ExportInput {
  diariaIds: number[];
}

export interface ExportResult {
  exported: number;
  integrationRef: string;
  exportedAt: string;
}

export interface DiariasPage {
  data: Diaria[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
}

export interface AuditLog {
  id: number;
  entityType: string;
  entityId: number;
  action: string;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  userName?: string | null;
  timestamp: string;
  /** @nullable */
  oldValues?: string | null;
  /** @nullable */
  newValues?: string | null;
}

export interface AuditPage {
  data: AuditLog[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
}

export interface DashboardSummary {
  pendentes: number;
  emAnalise: number;
  aprovadas: number;
  rejeitadas: number;
  disponivelExportacao: number;
  exportadas: number;
  pagas: number;
  canceladas: number;
  /** @nullable */
  totalValorAprovadas?: number | null;
  /** @nullable */
  totalValorExportadas?: number | null;
  /** @nullable */
  totalValorPagas?: number | null;
}

export interface TeamMetrics {
  teamId: number;
  teamName: string;
  total: number;
  pendentes: number;
  aprovadas: number;
  rejeitadas: number;
  exportadas: number;
  pagas: number;
  /** @nullable */
  totalValor?: number | null;
}

export interface ProviderMetrics {
  providerId: number;
  providerName: string;
  teamName: string;
  total: number;
  pendentes: number;
  aprovadas: number;
  exportadas: number;
  pagas: number;
  /** @nullable */
  totalValor?: number | null;
}

export interface ActivityItem {
  id: number;
  diariaId: number;
  providerName: string;
  teamName: string;
  action: string;
  userName: string;
  timestamp: string;
  status: string;
}

export type DiariasReportByStatus = {[key: string]: number};

export interface DiariasReport {
  data: Diaria[];
  totalRecords: number;
  /** @nullable */
  totalValor?: number | null;
  byStatus: DiariasReportByStatus;
  byTeam: TeamMetrics[];
}

export type ListProvidersParams = {
/**
 * @nullable
 */
teamId?: number | null;
};

export type ListDiariasParams = {
/**
 * @nullable
 */
status?: ListDiariasStatus;
/**
 * @nullable
 */
providerId?: number | null;
/**
 * @nullable
 */
teamId?: number | null;
/**
 * @nullable
 */
managerId?: number | null;
/**
 * @nullable
 */
startDate?: string | null;
/**
 * @nullable
 */
endDate?: string | null;
/**
 * @nullable
 */
page?: number | null;
/**
 * @nullable
 */
pageSize?: number | null;
};

export type ListDiariasStatus = typeof ListDiariasStatus[keyof typeof ListDiariasStatus] | null;


export const ListDiariasStatus = {
  pendente_aprovacao: 'pendente_aprovacao',
  em_analise: 'em_analise',
  aprovada: 'aprovada',
  rejeitada: 'rejeitada',
  solicitacao_correcao: 'solicitacao_correcao',
  disponivel_exportacao: 'disponivel_exportacao',
  exportada: 'exportada',
  paga: 'paga',
  cancelada: 'cancelada',
} as const;

export type ListAuditLogsParams = {
/**
 * @nullable
 */
entityType?: string | null;
/**
 * @nullable
 */
entityId?: number | null;
/**
 * @nullable
 */
userId?: number | null;
/**
 * @nullable
 */
startDate?: string | null;
/**
 * @nullable
 */
endDate?: string | null;
/**
 * @nullable
 */
page?: number | null;
/**
 * @nullable
 */
pageSize?: number | null;
};

export type GetReportDiariasParams = {
/**
 * @nullable
 */
status?: string | null;
/**
 * @nullable
 */
providerId?: number | null;
/**
 * @nullable
 */
teamId?: number | null;
/**
 * @nullable
 */
managerId?: number | null;
/**
 * @nullable
 */
startDate?: string | null;
/**
 * @nullable
 */
endDate?: string | null;
};