# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Controle de Diárias de Prestadores de Serviços
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /diarias:
    get:
      operationId: listDiarias
      tags: [diarias]
      summary: List daily rate records (filtered by role)
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum:
              - pendente_aprovacao
              - em_analise
              - aprovada
              - rejeitada
              - solicitacao_correcao
              - disponivel_exportacao
              - exportada
              - paga
              - cancelada
              - null
        - name: providerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: teamId
          in: query
          schema:
            type: ["integer", "null"]
        - name: managerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: startDate
          in: query
          schema:
            type: ["string", "null"]
            format: date
        - name: endDate
          in: query
          schema:
            type: ["string", "null"]
            format: date
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: pageSize
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Paginated daily rate records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiariasPage"

    post:
      operationId: createDiaria
      tags: [diarias]
      summary: Register a new daily rate (manager only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiariaInput"
      responses:
        "201":
          description: Created daily rate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Diaria"
  /diarias/{id}:
    get:
      operationId: getDiaria
      tags: [diarias]
      summary: Get a daily rate by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Daily rate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Diaria"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateDiaria
      tags: [diarias]
      summary: Update a pending daily rate (manager only, while pending)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiariaUpdate"
      responses:
        "200":
          description: Updated daily rate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Diaria"

    delete:
      operationId: deleteDiaria
      tags: [diarias]
      summary: Cancel a daily rate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Diaria"
  /diarias/{id}/approve:
    post:
      operationId: approveDiaria
      tags: [diarias]
      summary: Approve a daily rate (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiariaActionNote"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Diaria"
  /diarias/{id}/reject:
    post:
      operationId: rejectDiaria
      tags: [diarias]
      summary: Reject a daily rate (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiariaActionNote"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Diaria"
  /diarias/{id}/request-correction:
    post:
      operationId: requestCorrectionDiaria
      tags: [diarias]
      summary: Request correction on a daily rate (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiariaActionNote"
      responses:
        "200":
          description: Correction requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Diaria"
  /diarias/{id}/mark-paid:
    post:
      operationId: markDiariaPaid
      tags: [diarias]
      summary: Mark a daily rate as paid (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiariaActionNote"
      responses:
        "200":
          description: Marked as paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Diaria"
  /diarias/export:
    post:
      operationId: exportDiarias
      tags: [diarias]
      summary: Export approved daily rates to DECARGO People (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportInput"
      responses:
        "200":
          description: Export result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResult"

  # ── AUDIT ─────────────────────────────────────────────────────────────
...
  /reports/diarias:
    get:
      operationId: getReportDiarias
      tags: [reports]
      summary: Detailed report of daily rates with filters
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: providerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: teamId
          in: query
          schema:
            type: ["integer", "null"]
        - name: managerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: startDate
          in: query
          schema:
            type: ["string", "null"]
            format: date
        - name: endDate
          in: query
          schema:
            type: ["string", "null"]
            format: date
      responses:
        "200":
          description: Report data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiariasReport"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Diaria:
      type: object
      properties:
        id:
          type: integer
        providerId:
          type: integer
        providerName:
          type: string
        teamId:
          type: integer
        teamName:
          type: string
        managerId:
          type: ["integer", "null"]
        managerName:
          type: ["string", "null"]
        workDate:
          type: string
          format: date
        value:
          type: ["number", "null"]
          description: Null when viewer is a provider (no financial visibility)
        paymentDate:
          type: ["string", "null"]
          format: date
        observations:
          type: ["string", "null"]
        status:
          type: string
          enum:
            - pendente_aprovacao
            - em_analise
            - aprovada
            - rejeitada
            - solicitacao_correcao
            - disponivel_exportacao
            - exportada
            - paga
            - cancelada
        actionNote:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        createdByName:
          type: string
        approvedAt:
          type: ["string", "null"]
          format: date-time
        approvedByName:
          type: ["string", "null"]
        exportedAt:
          type: ["string", "null"]
          format: date-time
        exportedByName:
          type: ["string", "null"]
        integrationId:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
          format: date-time
        cancelledAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - providerId
        - providerName
        - teamId
        - teamName
        - workDate
        - status
        - createdAt
        - createdByName
    DiariaInput:
      type: object
      properties:
        providerId:
          type: integer
        teamId:
          type: integer
        workDate:
          type: string
          format: date
        value:
          type: number
        paymentDate:
          type: ["string", "null"]
          format: date
        observations:
          type: ["string", "null"]
      required: [providerId, teamId, workDate, value]
    DiariaUpdate:
      type: object
      properties:
        workDate:
          type: string
          format: date
        value:
          type: number
        paymentDate:
          type: ["string", "null"]
          format: date
        observations:
          type: ["string", "null"]
    DiariaActionNote:
      type: object
      properties:
        note:
          type: ["string", "null"]
    ExportInput:
      type: object
      properties:
        diariaIds:
          type: array
          items:
            type: integer
      required: [diariaIds]
    ExportResult:
      type: object
      properties:
        exported:
          type: integer
        integrationRef:
          type: string
        exportedAt:
          type: string
          format: date-time
      required: [exported, integrationRef, exportedAt]
    DiariasPage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Diaria"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, pageSize, totalPages]

    # ── AUDIT ─────────────────────────────────────────────────────────
...
    TeamMetrics:
      type: object
      properties:
        teamId:
          type: integer
        teamName:
          type: string
        total:
          type: integer
        pendentes:
          type: integer
        aprovadas:
          type: integer
        rejeitadas:
          type: integer
        exportadas:
          type: integer
        pagas:
          type: integer
        totalValor:
          type: ["number", "null"]
      required: [teamId, teamName, total, pendentes, aprovadas, rejeitadas, exportadas, pagas]
...
    DiariasReport:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Diaria"
        totalRecords:
          type: integer
        totalValor:
          type: ["number", "null"]
        byStatus:
          type: object
          additionalProperties:
            type: integer
        byTeam:
          type: array
          items:
            $ref: "#/components/schemas/TeamMetrics"
      required: [data, totalRecords, byStatus, byTeam]