# lib/api-zod/src/generated/types/diariaActionNote.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Controle de Diárias de Prestadores de Serviços
 * OpenAPI spec version: 0.1.0
 */

export interface DiariaActionNote {
  /** @nullable */
  note?: string | null;
}